Android port and project improvements
- Android port with GLES2 rendering and build infrastructure
- Application icons for Windows, Android, and installer
- Mobile UI: DPI scaling, drag-to-scroll, safe area insets, scrollable tab strips
- Audio fixes, shader precision fixes, logging improvements
- Licence change to PolyForm Noncommercial
- Git submodules for desktop builds instead of vcpkg

+++ source/Framework/Helpers.h
@@ -946,7 +946,7 @@ private:
 
 inline float RangedRandom(float min, float max)
 {
-    float f = (float) (rand()) / RAND_MAX;
+    float f = (float) (rand()) / (float) RAND_MAX;
     float res = min + f * (max - min);
     return res;
 }
